Cache plugin output by default if the user specifies CMS_PLUGIN_CACHE_DATA in settings. The default should be cachable, specifying which plugins shouldn't be cached explicitly.  The format of the setting is:
  CMS_PLUGIN_CACHE_DATA = {
    'PublicSearch': dict(cacheable=False),
    'CMSPollPlugin': dict(cacheable=False),

    # if you define CMS_PLUGIN_CACHE_DATA it must have a default key..
    'default': dict(cacheable=True, duration=CMS_CACHE_DURATION),
  }

The content caches uniquely on url/placeholder/plugin/language/user.

For anonymous users we will try to use the csrf_token as a way to uniquely identify the user.

For logged in users we use the username.

If we cannot find a way to uniquely identify a user we abandon caching.